lmms-eval/llava_repr_requirements.txt
llava@git+https://github.com/haotian-liu/LLaVA
accelerate>=0.21.0
black==24.1.0
datasets==2.16.1
evaluate>=0.4.0
jsonlines
numexpr
peft>=0.2.0
pybind11>=2.6.2
pytablewriter
rouge-score>=0.0.4
sacrebleu>=1.5.0
scikit-learn>=0.24.1
sqlitedict
openai>=1.0.0
pycocoevalcap
tqdm-multiprocess
transformers>=4.36.2
zstandard
pillow
pyyaml
sympy
mpmath
Jinja2
openpyxl
Levenshtein
hf_transfer
tenacity
wandb>=0.16.0
transformers-stream-generator
tiktoken
pre-commit